% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aide.R
\name{aide}
\alias{aide}
\alias{ave_sd}
\alias{average}
\alias{correlate}
\alias{iqr}
\alias{med}
\alias{med_iqr}
\alias{min_max}
\alias{stddev}
\alias{tertile}
\alias{trim_ws}
\title{aide, small and simple functions to help out.}
\usage{
med(x, digits = 1)

iqr(x, digits = 1)

correlate(x, y, digits = 2, method = "pearson", use = "complete.obs")

med_iqr(x, digits = 1)

average(x, digits = 1)

stddev(x, digits = 1)

ave_sd(x, digits = 1)

tertile(x)

trim_ws(x)

min_max(x, digits = 1)
}
\arguments{
\item{x}{Vector of values.}

\item{digits}{Number of decimal places for the output.}

\item{y}{Vector of values.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\description{
Simple utility functions that help with data analysis.
}

